+++ lib/trailblazer/activity/trace.rb
@@ -5,47 +5,63 @@ module Trailblazer
     #   stack, _ = Trailblazer::Activity::Trace.(activity, activity[:Start], { id: 1 })
     #   puts Trailblazer::Activity::Present.tree(stack) # renders the trail.
     #
-    # Hooks into the TaskWrap.
+    # Hooks into the taskWrap.
     module Trace
-      # {:argumenter} API
-      # FIXME: needs Introspect.arguments_for_call
-      # FIXME: needs TaskWrap.arguments_for_call
-      def self.arguments_for_call(activity, (options, flow_options), **circuit_options)
-        tracing_flow_options = {
-          stack:         Trace::Stack.new,
-        }
+      class << self
+        # {:argumenter} API
+        # FIXME: needs Introspect.arguments_for_call
+        # FIXME: needs TaskWrap.arguments_for_call
+        def arguments_for_call(activity, (options, flow_options), **circuit_options)
+          tracing_flow_options = {
+            stack:         Trace::Stack.new,
+          }
 
-        tracing_circuit_options = {
-          wrap_runtime:  ::Hash.new(Trace.wirings), # FIXME: this still overrides existing :wrap_runtime.
-        }
+          tracing_circuit_options = {
+            wrap_runtime:  ::Hash.new(Trace.wirings), # FIXME: this still overrides existing :wrap_runtime.
+          }
 
-        return activity, [ options, flow_options.merge(tracing_flow_options) ], circuit_options.merge(tracing_circuit_options)
-      end
+          return activity, [ options, tracing_flow_options.merge(flow_options) ], circuit_options.merge(tracing_circuit_options)
+        end
 
-      def self.call(activity, (options, flow_options), circuit_options={})
-        activity, (options, flow_options), circuit_options = Trace.arguments_for_call( activity, [options, flow_options], circuit_options ) # only run once for the entire circuit!
-        last_signal, (options, flow_options) =
-          Activity::TaskWrap.invoke(activity, [options, flow_options], circuit_options)
+        def call(activity, (options, flow_options), circuit_options={})
+          activity, (options, flow_options), circuit_options = Trace.arguments_for_call( activity, [options, flow_options], circuit_options ) # only run once for the entire circuit!
 
-        return flow_options[:stack].to_a, last_signal, [options, flow_options]
-      end
+          last_signal, (options, flow_options) =
+            Activity::TaskWrap.invoke(activity, [options, flow_options], circuit_options)
 
-      private
+          return flow_options[:stack], last_signal, [options, flow_options]
+        end
 
-      # Insertions for the trace tasks that capture the arguments just before calling the task,
-      # and before the TaskWrap is finished.
-      #
-      # Note that the TaskWrap steps are implemented in Activity::TaskWrap::Trace.
-      def self.wirings
-        Module.new do
-          extend Activity::Path::Plan()
+        alias_method :invoke, :call
 
-          task TaskWrap::Trace.method(:capture_args),   id: "task_wrap.capture_args",   before: "task_wrap.call_task"
-          task TaskWrap::Trace.method(:capture_return), id: "task_wrap.capture_return", before: "End.success", group: :end
+        # Insertions for the trace tasks that capture the arguments just before calling the task,
+        # and before the TaskWrap is finished.
+        #
+        # Note that the TaskWrap steps are implemented in Activity::TaskWrap::Trace.
+        #
+        # @private
+        def wirings
+          Module.new do
+            extend Activity::Path::Plan()
+
+            task TaskWrap::Trace.method(:capture_args),   id: "task_wrap.capture_args",   before: "task_wrap.call_task"
+            task TaskWrap::Trace.method(:capture_return), id: "task_wrap.capture_return", before: "End.success", group: :end
+          end
         end
       end
 
-      Entity = Struct.new(:task, :activity, :type, :value, :value2)
+      Entity = Struct.new(:task, :activity, :data)
+      class Entity::Input < Entity
+      end
+
+      class Entity::Output < Entity
+      end
+
+      class Task < Array
+        def inspect
+          %{<Task>#{super}}
+        end
+      end
 
       # Mutable/stateful per design. We want a (global) stack!
       class Stack
@@ -55,7 +71,7 @@ module Trailblazer
         end
 
         def indent!
-          current << indented = []
+          current << indented = Task.new
           @stack << indented
         end